Move dungeon map hit ref grammar to runtime
