+++ src/view/slotcontent/main/dungeonmap/DungeonMapPreviewDiffContentPartModel.java
@@ -6,6 +6,7 @@ import src.domain.dungeon.published.DungeonCellRef;
 import src.domain.dungeon.published.DungeonEdgeRef;
 import src.domain.dungeon.published.DungeonEditorStateSnapshot;
 import src.domain.dungeon.published.DungeonFeatureKind;
+import src.features.dungeon.runtime.DungeonEditorMapHitRefs;
 import src.features.dungeon.runtime.DungeonEditorPreparedFrameFacts.PreviewAreaDiffFrame;
 import src.features.dungeon.runtime.DungeonEditorPreparedFrameFacts.PreviewBoundaryDiffFrame;
 import src.features.dungeon.runtime.DungeonEditorPreparedFrameFacts.PreviewFeatureDiffFrame;
@@ -111,7 +112,7 @@ final class DungeonMapPreviewDiffContentPartModel {
             DungeonMapContentModel.MapInteractionFrame interactionFrame
     ) {
         return interactionFrame.previewHandleHitRefs()
-                .contains(src.domain.dungeon.published.DungeonEditorMapHitRef.marker(handle.ref(), handle.cell()).value());
+                .contains(DungeonEditorMapHitRefs.marker(handle.ref(), handle.cell()).value());
     }
 
     private Marker previewHandleMarker(
